requirements.txt
pandas
pandasai
langchain
python-dotenv
chainlit